+++ src/test/java/CreateCourierParametrizedTest.java
@@ -1,7 +1,6 @@
-import couriers.*;
 import io.qameta.allure.Description;
-import io.restassured.RestAssured;
 import io.restassured.response.Response;
+import org.example.couriers.*;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -10,6 +9,7 @@ import org.junit.runners.Parameterized;
 @RunWith(Parameterized.class)
 public class CreateCourierParametrizedTest {
     CourierMethods method = new CourierMethods();
+    Courier courier;
     Response response;
     private final String login;
     private final String password;
@@ -31,14 +31,12 @@ public class CreateCourierParametrizedTest {
 
     @Before
     public void setUp() {
-        RestAssured.baseURI = "https://qa-scooter.praktikum-services.ru";
+        courier = new Courier(login, password, "Tasha");
     }
 
     @Test
     @Description("1.6 если одного из полей нет, запрос возвращает ошибку")
     public void createCourierMissingRequiredFieldTest() {
-        Courier courier = new Courier(login, password, "Tasha");
-
         response = method.requestCreateCourier(courier);
         method.responseCreateCourierErrorMissing(response);
     }
