+++ src/test/java/CreateCourierTest.java
@@ -1,7 +1,6 @@
-import couriers.*;
 import io.qameta.allure.Description;
-import io.restassured.RestAssured;
 import io.restassured.response.Response;
+import org.example.couriers.*;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -9,14 +8,15 @@ import org.junit.Test;
 public class CreateCourierTest {
     CourierMethods method = new CourierMethods();
     CourierGen generator = new CourierGen();
-    Courier courier = generator.random();
-    Credentials cred = Credentials.from(courier);
+    Courier courier;
+    Credentials cred;
     Response response;
     int courierId;
 
     @Before
     public void setUp() {
-        RestAssured.baseURI = "https://qa-scooter.praktikum-services.ru";
+        courier = generator.random();
+        cred = Credentials.from(courier);
 
         response = method.requestAuthCourier(cred);                         //Проверка может есть такой и дел если что
         if (response.statusCode() == 200) {
